feat(webapp): new expense view
