+++ webapp/src/gql/index.js
@@ -1,5 +1,5 @@
 export * from './general.gql'
 export * from './income.gql'
+export * from './delegate2.gql'
 export * from './expense.gql'
-export * from './expenseGeneral.gql'
 export * from './delegate.gql'
